Fixed some debug prints in windows.

+++ src/windows/drop.cpp
@@ -82,7 +82,7 @@ bool FileDropTarget::QueryDataObject(IDataObject *pDataObject) {
   HRESULT r;
 
   r = pDataObject->QueryGetData(&fmtetc);
-  debug(DEBUG_TRACE, "WIN32", "QueryGetData: 0x%08X", r);
+  debug(DEBUG_TRACE, "WIN32", "QueryGetData: 0x%08X", (uint32_t)r);
 
   return SUCCEEDED(r);
 }
@@ -145,11 +145,11 @@ bool FileDropTarget::DropData(HWND hwnd, IDataObject *pDataObject) {
   bool ok = false;
 
   r = pDataObject->QueryGetData(&fmtetc);
-  debug(DEBUG_TRACE, "WIN32", "DropData QueryGetData: 0x%08X", r);
+  debug(DEBUG_TRACE, "WIN32", "DropData QueryGetData: 0x%08X", (uint32_t)r);
   if (SUCCEEDED(r)) {
     debug(DEBUG_TRACE, "WIN32", "DropData query data ok");
     r = pDataObject->GetData(&fmtetc, &stgmed);
-    debug(DEBUG_TRACE, "WIN32", "DropData GetData: 0x%08X", r);
+    debug(DEBUG_TRACE, "WIN32", "DropData GetData: 0x%08X", (uint32_t)r);
     if (SUCCEEDED(r)) {
       debug(DEBUG_TRACE, "WIN32", "DropData get data ok");
       hDrop = (HDROP)stgmed.hGlobal;
